requirements.txt
smbus2
RPi.GPIO
